+++ src/m_config.h
@@ -94,9 +94,7 @@ typedef enum
 
 #define am_grid_default                         false
 
-#define am_gridsize_min                         4
-#define am_gridsize_default                     128
-#define am_gridsize_max                         4096
+#define am_gridsize_default                     "128x128"
 
 #define am_gridcolor_min                        0
 #define am_gridcolor_default                    5